+++ src/main/kotlin/operator/ShopCalculator.kt
@@ -2,29 +2,6 @@ package org.example.operator
 
 import org.example.product.Product
 
-class ShopCalculator {
-
-    fun calculateTotal(products: List<Product>): Double {
-        var total = 0.0
-        products.forEach {
-            p -> total += calculatePrice(p)
-        }
-        return total
-    }
-
-    private fun calculatePrice(product: Product): Double {
-        val isUnit = product.price.isUnit
-        return when {
-            isUnit -> calculatePriceByUnit(product)
-            else -> calculatePriceByWeight(product)
-        }
-    }
-
-    private fun calculatePriceByWeight(product: Product): Double {
-        return product.price.value * product.price.qty!!
-    }
-
-    private fun calculatePriceByUnit(product: Product): Double {
-        return product.price.value
-    }
+interface ShopCalculator {
+    fun calculateTotal(products: List<Product>): Double
 }
